chore: remove service worker extension runtime
