+++ app/client/src/components/header/header.tsx
@@ -5,7 +5,6 @@ import {
   selectedExtensionState,
 } from "../../states/extensions.ts";
 import { For, onMount } from "solid-js";
-import { loadExtensionWorker } from "../../extensionWorker.ts";
 import { createTakos } from "../../takos.ts";
 
 export default function ChatHeader() {
@@ -32,8 +31,7 @@ export default function ChatHeader() {
         const list = data[0]?.result ?? [];
         setExtensions(list);
         for (const ext of list) {
-          const takos = createTakos(ext.identifier);
-          loadExtensionWorker(ext.identifier, takos).catch(() => {});
+          createTakos(ext.identifier);
         }
       }
     } catch (_e) {
